test(example, ios): quiet non-module warnings in ios build
Flipper has the noisiest impact on build logs I have seen in a long time, and obscures
warnings we really want to see in our module. This stance in the Podfile inhibits warnings
when they are not for our module. I tested by removing the deprecation fix a couple commits ago
and rebuilding, and I verified I still saw the deprecation warning in our module even with
this change in place.